chore(binary): removed subpackage 'types' for data types
