+++ src/main/java/marshmalliow/core/binary/data/RegisteredDataType.java
@@ -1,0 +1,4 @@
+package marshmalliow.core.binary.data;
+
+public interface RegisteredDataType {
+}
