+++ src/main/java/marshmalliow/core/binary/data/advanced/StringDataType.java
@@ -1,8 +1,8 @@
-package marshmalliow.core.binary.data.types.primitive;
+package marshmalliow.core.binary.data.advanced;
 
 import java.io.IOException;
 
-import marshmalliow.core.binary.data.types.DataType;
+import marshmalliow.core.binary.data.DataType;
 import marshmalliow.core.binary.registry.DataTypeEnum;
 import marshmalliow.core.binary.registry.DataTypeRegistry;
 import marshmalliow.core.binary.utils.Charset;
@@ -10,42 +10,38 @@ import marshmalliow.core.io.BinaryReader;
 import marshmalliow.core.io.BinaryWriter;
 
 /**
- * Boolean data type corresponding to the format specification.
+ * String data type corresponding to the format specification.
  * @version	1.0.0
  * @author 278deco
  */
-public class BooleanDataType extends DataType<Boolean> {
+public class StringDataType extends DataType<String> {
 
-	public BooleanDataType() {
+	public StringDataType() {
 		super();
 	}
 
-	public BooleanDataType(String name, boolean value) {
+	public StringDataType(String name, String value) {
 		super(name, value);
 	}
-	
-	public BooleanDataType(boolean value) {
-		super(value);
-	}
-	
+
 	@Override
 	public void writeValue(BinaryWriter writer, DataTypeRegistry registry, Charset charset) throws IOException {
-		writer.writeBoolean(this.getValue());
+		writer.writeUTF(this.getValue(), charset);
 	}
 
 	@Override
 	public void readValue(BinaryReader reader, DataTypeRegistry registry, Charset charset) throws IOException {
-		this.setValue(reader.readBoolean());
+		this.setValue(reader.readUTF(charset));
 	}
 
 	@Override
 	public byte getId() {
-		return DataTypeEnum.BOOLEAN.getId();
+		return DataTypeEnum.STRING.getId();
 	}
 	
 	@Override
 	public Category getCategory() {
-		return Category.PRIMITIVE;
+		return Category.ADVANCED;
 	}
 
 }
