+++ src/main/java/marshmalliow/core/binary/data/container/ArrayDataType.java
@@ -1,10 +1,10 @@
-package marshmalliow.core.binary.data.types.container;
+package marshmalliow.core.binary.data.container;
 
 import java.io.IOException;
 import java.lang.reflect.Constructor;
 import java.util.Arrays;
 
-import marshmalliow.core.binary.data.types.DataType;
+import marshmalliow.core.binary.data.DataType;
 import marshmalliow.core.binary.registry.DataTypeEnum;
 import marshmalliow.core.binary.registry.DataTypeRegistry;
 import marshmalliow.core.binary.utils.Charset;
