+++ src/main/java/marshmalliow/core/binary/data/primitive/BooleanDataType.java
@@ -1,8 +1,8 @@
-package marshmalliow.core.binary.data.types.primitive;
+package marshmalliow.core.binary.data.primitive;
 
 import java.io.IOException;
 
-import marshmalliow.core.binary.data.types.DataType;
+import marshmalliow.core.binary.data.DataType;
 import marshmalliow.core.binary.registry.DataTypeEnum;
 import marshmalliow.core.binary.registry.DataTypeRegistry;
 import marshmalliow.core.binary.utils.Charset;
@@ -10,37 +10,37 @@ import marshmalliow.core.io.BinaryReader;
 import marshmalliow.core.io.BinaryWriter;
 
 /**
- * Integer data type corresponding to the format specification.
+ * Boolean data type corresponding to the format specification.
  * @version	1.0.0
  * @author 278deco
  */
-public class IntegerDataType extends DataType<Integer> {
+public class BooleanDataType extends DataType<Boolean> {
 
-	public IntegerDataType() {
+	public BooleanDataType() {
 		super();
 	}
 
-	public IntegerDataType(String name, int value) {
+	public BooleanDataType(String name, boolean value) {
 		super(name, value);
 	}
 	
-	public IntegerDataType(int value) {
+	public BooleanDataType(boolean value) {
 		super(value);
 	}
-
+	
 	@Override
 	public void writeValue(BinaryWriter writer, DataTypeRegistry registry, Charset charset) throws IOException {
-		writer.writeInt(this.getValue());
+		writer.writeBoolean(this.getValue());
 	}
 
 	@Override
 	public void readValue(BinaryReader reader, DataTypeRegistry registry, Charset charset) throws IOException {
-		this.setValue(reader.readInt());
+		this.setValue(reader.readBoolean());
 	}
 
 	@Override
 	public byte getId() {
-		return DataTypeEnum.INT.getId();
+		return DataTypeEnum.BOOLEAN.getId();
 	}
 	
 	@Override
