+++ src/main/java/marshmalliow/core/binary/data/primitive/ByteDataType.java
@@ -1,8 +1,8 @@
-package marshmalliow.core.binary.data.types.primitive;
+package marshmalliow.core.binary.data.primitive;
 
 import java.io.IOException;
 
-import marshmalliow.core.binary.data.types.DataType;
+import marshmalliow.core.binary.data.DataType;
 import marshmalliow.core.binary.registry.DataTypeEnum;
 import marshmalliow.core.binary.registry.DataTypeRegistry;
 import marshmalliow.core.binary.utils.Charset;
@@ -10,42 +10,49 @@ import marshmalliow.core.io.BinaryReader;
 import marshmalliow.core.io.BinaryWriter;
 
 /**
- * Long data type corresponding to the format specification.
+ * Byte data type corresponding to the format specification.
  * @version	1.0.0
  * @author 278deco
  */
-public class LongDataType extends DataType<Long> {
+public class ByteDataType extends DataType<Byte> {
 
-	public LongDataType() {
+	public ByteDataType() {
 		super();
 	}
 
-	public LongDataType(String name, long value) {
+	public ByteDataType(String name, byte value) {
 		super(name, value);
 	}
 	
-	public LongDataType(long value) {
+	public ByteDataType(byte value) {
 		super(value);
+	}
+	
+	public static ByteDataType[] asArray(byte...values) {
+		final ByteDataType[] result = new ByteDataType[values.length];
+		for(int i = 0; i < result.length; i++) result[i] = new ByteDataType(values[i]);
+		
+		return result;
 	}
 
 	@Override
 	public void writeValue(BinaryWriter writer, DataTypeRegistry registry, Charset charset) throws IOException {
-		writer.writeLong(this.getValue());
+		writer.writeByte(this.getValue());
 	}
 
 	@Override
 	public void readValue(BinaryReader reader, DataTypeRegistry registry, Charset charset) throws IOException {
-		this.setValue(reader.readLong());
+		this.setValue(reader.readByte());
 	}
 
 	@Override
 	public byte getId() {
-		return DataTypeEnum.LONG.getId();
+		return DataTypeEnum.BYTE.getId();
 	}
-	
+
 	@Override
 	public Category getCategory() {
 		return Category.PRIMITIVE;
 	}
-
+	
 }
