--- src/main/java/marshmalliow/core/binary/data/types/RegisteredDataType.java
@@ -1,4 +1,0 @@
-package marshmalliow.core.binary.data.types;
-
-public interface RegisteredDataType {
-}
