+++ src/main/java/marshmalliow/core/binary/registry/DataTypeRegistry.java
@@ -3,8 +3,8 @@ package marshmalliow.core.binary.registry;
 import java.util.HashMap;
 import java.util.Map;
 
-import marshmalliow.core.binary.data.types.DataType;
-import marshmalliow.core.binary.data.types.RegisteredDataType;
+import marshmalliow.core.binary.data.DataType;
+import marshmalliow.core.binary.data.RegisteredDataType;
 import marshmalliow.core.exceptions.DatatypeRegistryException;
 
 public class DataTypeRegistry {
